requirements.txt
# --- Core numerics & ML (Phase 1 baselines + general tooling) ---
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# --- Deep learning & NLP (later phases / transformers) ---
torch>=2.0.0
transformers>=4.40.0
sentencepiece>=0.1.99
accelerate>=0.27.0
peft>=0.10.0

# --- Graph / probabilistic reasoning ---
networkx==3.2.1
pgmpy==0.1.25

# --- Documents & OCR ---
python-docx>=1.1.0
pdfplumber==0.10.3
Pillow==10.2.0
opencv-python>=4.9.0.80
pytesseract>=0.3.10

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.20.0

# --- Notebooks (Jupyter + IPython.display in notebooks) ---
jupyterlab>=4.0.0
ipykernel>=6.29.0
ipython>=8.12.0
notebook>=7.0.0

# --- CUAD / Excel label reports (optional EDA on .xlsx) ---
openpyxl>=3.1.0

# --- Phase 2 (configs + interpretability) ---
PyYAML>=6.0.0
shap>=0.44.0

# --- Phase 3 (hybrid reasoning + app) ---
streamlit==1.32.0
pdf2image==1.17.0
easyocr==1.7.1
